+++ LINALG/include/vector.h
@@ -6,8 +6,8 @@ class vector
 public:
 	void add(const vector* v);
 	void subtract(const vector* v);
-	void multiply(int scalar);
+	void multiply(float scalar);
 	void render(SDL_Renderer* renderer_, int k_screen_width, int k_screen_height);
-	int x;
-	int y;
+	float x;
+	float y;
 };
